+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Omit Former holdings ID when copying holdings record. Refs UIIN-845.
 * Filter instance by "Call number, eye readable" via holdings segment. Refs UIIN-858.
 * Filter instance by "Call number, eye readable" via items segment. Refs UIIN-990.
+* Enhance preceding connected titles. Refs UIIN-960.
 
 ## [1.13.1](https://github.com/folio-org/ui-inventory/tree/v1.13.1) (2019-12-11)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.13.0...v1.13.1)